Fixed issues that Armine addressed
