+++ javinukai-front/src/Components/user-management/PaginationSettings.jsx
@@ -5,22 +5,9 @@ import { useTranslation } from "react-i18next";
 export default function PaginationSettings({
   pagination,
   setPagination,
-  availablePageNumber,
 }) {
   const [searchedSurname, setSearchedSurname] = useState("");
 
-  function handleNextPage() {
-    if (
-      availablePageNumber === 1 ||
-      pagination.page === availablePageNumber - 1
-    )
-      return;
-    setPagination({ ...pagination, page: pagination.page + 1 });
-  }
-  function handlePreviousPage() {
-    if (pagination.page === 0) return;
-    setPagination({ ...pagination, page: pagination.page - 1 });
-  }
   function handleLimitChange(e) {
     e.preventDefault();
     setPagination({ ...pagination, limit: e.target.value, page: 0 });
@@ -47,26 +34,9 @@ export default function PaginationSettings({
 
   const { t } = useTranslation();
   return (
-    <div className="flex flex-col px-2 py-3 space-y-2 xl:grid xl:grid-cols-4 xl:grid-rows-2 xl:gap-2">
-      <section className="flex space-x-3 justify-center xl:col-span-4 xl:h-fit xl:self-end">
-        <Button disabled={pagination.page == 0} onClick={handlePreviousPage}>
-          {t("PaginationSettings.previousPage")}
-        </Button>
-        <p className="text-xl py-1 px-4 rounded-full bg-white">
-          {pagination.page + 1 + "/" + (availablePageNumber || pagination.page)}
-        </p>
-        <Button
-          disabled={
-            availablePageNumber === 1 ||
-            pagination.page === availablePageNumber - 1
-          }
-          extraStyle="px-8"
-          onClick={handleNextPage}
-        >
-          {t("PaginationSettings.nextPage")}
-        </Button>
-      </section>
-      <section className="flex flex-col space-y-2">
+    <div className="flex flex-col px-2 py-5 space-y-2 xl:grid xl:grid-cols-4 xl:gap-2">
+      
+      <section className="flex flex-col space-y-2 py-3">
         <label>{t("PaginationSettings.usersTitle")}</label>
         <select
           className="py-2 px-4 bg-white rounded-md"
@@ -105,7 +75,7 @@ export default function PaginationSettings({
             {t("PaginationSettings.fieldIsEnabled")}
           </option>
           <option value="isNonLocked">
-            {t("PaginationSettings.fieldIsNonLocked")}
+            {t("PaginationSettings.fieldIsNonLocked")}         
           </option>
         </select>
       </section>
